:recycle: refactored code and optimized file structure for better readability and maintainability
