+++ src/constants/client.constant.ts
@@ -1,0 +1,1 @@
+export const CLIENT_ID = '383226320970055681' as const;
