+++ src/constants/empty.constant.ts
@@ -1,0 +1,2 @@
+export const EMPTY = '' as const;
+export const FAKE_EMPTY = '\u200b\u200b' as const;
