+++ src/constants/git.constant.ts
@@ -1,0 +1,2 @@
+export const UNKNOWN_GIT_BRANCH = 'Unknown' as const;
+export const UNKNOWN_GIT_REPO_NAME = 'Unknown' as const;
