+++ src/constants/languages.constant.ts
@@ -1,0 +1,4 @@
+import LANG from '../data/languages.json';
+
+export const KNOWN_EXTENSIONS: { [key: string]: { image: string } } = LANG.KNOWN_EXTENSIONS;
+export const KNOWN_LANGUAGES: { language: string; image: string }[] = LANG.KNOWN_LANGUAGES;
