+++ src/constants/logLevel.constant.ts
@@ -1,0 +1,7 @@
+export const enum LogLevel {
+    Trace = 'TRACE',
+    Debug = 'DEBUG',
+    Info = 'INFO',
+    Warn = 'WARN',
+    Error = 'ERROR',
+}
