+++ src/constants/sizes.constant.ts
@@ -1,0 +1,1 @@
+export const FILE_SIZES = [' bytes', 'KB', 'MB', 'GB', 'TB'] as const;
